Revert "- Renamed mod to "enchantment-text-color""
This reverts commit 6e380815e3aef999996a0c64e88768cf831ec1db.
