+++ gradle.properties
@@ -3,8 +3,8 @@ org.gradle.jvmargs=-Xmx1G
 org.gradle.parallel=true
 
 mod_version=0.1.0-alpha
-maven_group=com.shadowhunter22.enchantmenttextcolor
-archives_base_name=enchantment-text-color
+maven_group=com.shadowhunter22.coloredenchantments
+archives_base_name=colored-enchantments
 
 minecraft_version=1.21
 yarn_mappings=1.21+build.9
