+++ src/client/resources/colored-enchantments.client.mixins.json
@@ -1,0 +1,10 @@
+{
+	"required": true,
+	"package": "dev.shadowhunter22.coloredenchantments.mixin.client",
+	"compatibilityLevel": "JAVA_21",
+	"client": [
+	],
+	"injectors": {
+		"defaultRequire": 1
+	}
+}
