+++ src/main/java/dev/shadowhunter22/coloredenchantments/api/registry/ModRegistryKeys.java
@@ -3,10 +3,10 @@
 // See LICENSE file in the project root for details.
 //
 
-package dev.shadowhunter22.enchantmenttextcolor.api.registry;
+package dev.shadowhunter22.coloredenchantments.api.registry;
 
-import dev.shadowhunter22.enchantmenttextcolor.EnchantmentTextColor;
-import dev.shadowhunter22.enchantmenttextcolor.api.EnchantmentStyling;
+import dev.shadowhunter22.coloredenchantments.ColoredEnchantments;
+import dev.shadowhunter22.coloredenchantments.api.EnchantmentStyling;
 import net.minecraft.registry.Registry;
 import net.minecraft.registry.RegistryKey;
 import net.minecraft.util.Identifier;
@@ -15,6 +15,6 @@ public class ModRegistryKeys {
     public static final RegistryKey<Registry<EnchantmentStyling>> STYLING = of("styling");
 
     private static <T> RegistryKey<Registry<T>> of(String id) {
-        return RegistryKey.ofRegistry(Identifier.of(EnchantmentTextColor.MOD_ID, id));
+        return RegistryKey.ofRegistry(Identifier.of(ColoredEnchantments.MOD_ID, id));
     }
 }
