+++ src/main/resources/assets/colored-enchantments/icon.png
