+++ src/testmod/java/dev/shadowhunter22/coloredenchantments/test/ColoredEnchantmentsTestmodDataGen.java
@@ -1,4 +1,4 @@
-package dev.shadowhunter22.enchantmenttextcolor.test;//
+package dev.shadowhunter22.coloredenchantments.test;//
 // Copyright (c) 2024 by ShadowHunter22. All rights reserved.
 // See LICENSE file in the project root for details.
 //
