+++ src/testmod/java/dev/shadowhunter22/coloredenchantments/test/EnchantmentTextData.java
@@ -1,9 +1,9 @@
-package dev.shadowhunter22.enchantmenttextcolor.test;//
+package dev.shadowhunter22.coloredenchantments.test;//
 // Copyright (c) 2024 by ShadowHunter22. All rights reserved.
 // See LICENSE file in the project root for details.
 //
 
-import dev.shadowhunter22.enchantmenttextcolor.api.datagen.v1.EnchantmentTextProvider;
+import dev.shadowhunter22.coloredenchantments.api.datagen.v1.EnchantmentTextProvider;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.minecraft.enchantment.Enchantments;
 import net.minecraft.registry.RegistryWrapper;
@@ -20,7 +20,7 @@ public class EnchantmentTextData extends EnchantmentTextProvider {
 	public void generate() {
 		// testing to see if the color will be red when adding .color() to the method chain
 		this.addEntry(Enchantments.AQUA_AFFINITY)
-				.color(Formatting.GREEN.getColorValue())
+				.color(Formatting.RED.getColorValue())
 				.build();
 
 		// testing to see if the color will still be red without adding .color() to the method chain
